Move actions into the protocol
